client/requirements.txt
PyYAML==3.11
certifi==14.05.14
greenlet==0.4.4
msgpack-python==0.4.2
msgpack-rpc-python==0.3.3
neovim==0.0.22
tornado==4.0.2
